requirements.txt
# Dependências do Sistema de Cálculo de Rentabilidade
# Versões mínimas recomendadas para cada pacote

# Framework web para interface do usuário
streamlit>=1.28.0

# Manipulação e análise de dados
pandas>=2.0.0
numpy>=1.24.0

# Requisições HTTP para APIs
requests>=2.31.0

# Visualização de dados
matplotlib>=3.7.0
seaborn>=0.12.0

# Integração com Yahoo Finance
yfinance>=0.2.36

# Utilitários
python-dateutil>=2.8.2  # Manipulação de datas
xlsxwriter>=3.1.0       # Exportação para Excel
jinja2>=3.0.0          # Templates para exportação

# Novas dependências adicionadas
plotly>=5.18.0
pytz>=2023.3
openpyxl>=3.1.2
xlrd>=2.0.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
scipy>=1.11.0
statsmodels>=0.14.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.6.0
